+++ src/main/java/lotto/domain/YieldRate.java
@@ -3,7 +3,7 @@ package lotto.domain;
 public class YieldRate {
     private double yieldRate;
 
-    public YieldRate(int spendMoney, int gainMoney) {
+    public YieldRate(int spendMoney, long gainMoney) {
         this.yieldRate = (double) gainMoney / spendMoney;
     }
